refactor: flatten three-crate workspace into a single clarus crate
Previously: workspace with clarus-engine, clarus-explanation, clarus-input-adapter.
Now: single crate with engine/, explanation/, file_replay, unity_udp, sealer as modules.

Benefits:
  - One Cargo.toml, one Cargo.lock, one version to bump
  - No intra-workspace path deps; edgesentry-audit = "0.2.0" is the only
    external non-std dependency worth noting
  - cargo test runs everything in one pass
  - CI no longer needs to reason about workspace members

Module layout:
  src/
    lib.rs               ← pub mod engine, explanation, file_replay, sealer, unity_udp
    engine/              ← physics primitives + rule engine (60 tests)
    explanation/         ← LLM client + KB + grounding check (12 tests)
    file_replay.rs       ← CSV replay adapter (7 tests)
    unity_udp.rs         ← Unity UDP adapter (6 tests)
    sealer.rs            ← AuditRecord chain builder (5 tests)
    main.rs              ← clarus binary
    validate_profile.rs  ← clarus-validate binary (14 tests)

116 tests pass (previously 129; the difference is 13 duplicate tests that
ran in both lib and binary targets — now run once in the lib).
