+++ Cargo.toml
@@ -1,7 +1,22 @@
-[workspace]
-members = [
-    "crates/engine",
-    "crates/input-adapter",
-    "crates/explanation",
-]
-resolver = "2"
+[package]
+name = "clarus"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+edgesentry-audit = "0.2.0"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+ureq = { version = "2", features = ["json"] }
+
+[dev-dependencies]
+tempfile = "3"
+
+[[bin]]
+name = "clarus"
+path = "src/main.rs"
+
+[[bin]]
+name = "clarus-validate"
+path = "src/validate_profile.rs"
