--- crates/engine/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "clarus-engine"
-version = "0.1.0"
-edition = "2021"
-description = "Physics-based risk event engine for clarus"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
