--- crates/explanation/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "clarus-explanation"
-version = "0.1.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-clarus-engine = { path = "../engine" }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-ureq = { version = "2", features = ["json"] }
