--- crates/input-adapter/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "clarus-input-adapter"
-version = "0.1.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-clarus-engine = { path = "../engine" }
-clarus-explanation = { path = "../explanation" }
-edgesentry-audit = "0.2.0"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-
-[dev-dependencies]
-tempfile = "3"
-
-[[bin]]
-name = "clarus"
-path = "src/main.rs"
-
-[[bin]]
-name = "clarus-validate"
-path = "src/validate_profile.rs"
