--- crates/input-adapter/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod file_replay;
-pub mod unity_udp;
