+++ src/explanation/kb.rs
